Refactor current tests to integration and made new unit test to be written for _check_and_return_valid_data_format, but it is empty
